Remove old crawling system
