+++ Content.Goobstation.Shared/Knockdown/KnockdownOnCollideSystem.cs
@@ -8,7 +8,7 @@
 
 using Content.Shared._Goobstation.Heretic.Components;
 using Content.Shared._Goobstation.Wizard.Mutate;
-using Content.Shared._White.Standing;
+using Content.Shared.Standing;
 using Content.Shared.Projectiles;
 using Content.Shared.Throwing;
 
@@ -16,7 +16,7 @@ namespace Content.Goobstation.Shared.Knockdown;
 
 public sealed class KnockdownOnCollideSystem : EntitySystem
 {
-    [Dependency] private readonly SharedLayingDownSystem _layingDown = default!;
+    [Dependency] private readonly StandingStateSystem _standing = default!;
     [Dependency] private readonly SharedHulkSystem _hulk = default!;
 
     public override void Initialize()
@@ -48,6 +48,6 @@ public sealed class KnockdownOnCollideSystem : EntitySystem
         if (HasComp<RustbringerComponent>(target))
             return;
 
-        _layingDown.TryLieDown(target, null, null, component.Behavior);
+        _standing.Down(target);
     }
 }
